Migrate usage of histogramming class to VertexAnalyzer
