+++ TMTrackTrigger/l1VertexFinder/python/VertexAnalyzer_cff.py
@@ -1,3 +1,8 @@
 import FWCore.ParameterSet.Config as cms
 
-L1TVertexAnalyzer = cms.EDAnalyzer('VertexAnalyzer')
+L1TVertexAnalyzer = cms.EDAnalyzer('VertexAnalyzer',
+  tpInputTag = cms.InputTag("mix", "MergedTrackTruth"),
+  stubInputTag = cms.InputTag("TTStubsFromPhase2TrackerDigis", "StubAccepted"),
+  stubTruthInputTag = cms.InputTag("TTStubAssociatorFromPixelDigis", "StubAccepted"),
+  clusterTruthInputTag = cms.InputTag("TTClusterAssociatorFromPixelDigis", "ClusterAccepted"),
+)
